+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 3.2.0 (July 5, 2021)
 * createObject method now supports up to 5 query headers
-* rename lookupObjectByQueryParameter method to lookupObjectByQueryParameters
-* lookupObjectByQueryParameters method now now supports up to 5 query parameters
+* rename lookupObjectByQueryParameter method to lookupObjectsByQueryParameters
+* lookupObjectsByQueryParameters method now now supports up to 5 query parameters
 
 # 3.1.0 (July 1, 2021)
 * lookupObjectByQueryParameter method now returns JSON array instead of its string representation